Read data from a Felt map

Read simple features from a Felt map or get data embedded in the website of a Felt map. Superseded by read_felt_map().

Usage

read_felt(
  url,
  map_id = NULL,
  ...,
  crs = 3857,
  token = NULL,
  rename = TRUE,
  name_repair = "check_unique"
)

Arguments

map_id

A Felt map URL, map ID string, or a named list with a id and type element. If map_id is a list, it must be equivalent to the output from get_felt_map() where the list includes a "id" string and a "type" string with the value "map".

...

Additional parameters passed to sf::read_sf().

crs

Coordinate reference system to return. Defaults to 3857.

token

Felt personal access token

rename

If TRUE (default), strip the prefix text "felt-" from all column names.

name_repair

Passed to repair parameter of vctrs::vec_as_names(). Defaults to "check_unique".

Value

A simple feature data frame.

See also