Set or get a Felt API personal access token

An API personal access token is required to use read_felt(). See https://feltmaps.notion.site/Felt-Public-API-reference-PUBLIC-c01e0e6b0d954a678c608131b894e8e1 for instructions on how to get a token.

Usage

set_felt_token(
  token = NULL,
  install = FALSE,
  overwrite = FALSE,
  default = "FELT_ACCESS_TOKEN"
)

get_felt_token(
  token = NULL,
  default = "FELT_ACCESS_TOKEN",
  call = caller_env()
)

Arguments

token

Felt personal access token

install

If TRUE, this function adds your token to your .Renviron for use in future sessions. Defaults to FALSE.

overwrite

If TRUE, overwrite any existing token in .Renviron using the same environment variable name. Defaults to FALSE.

default

Default name used for environment variable where the token is saved.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.